requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
pdfplumber==0.11.0
pypdf==5.1.0
Werkzeug==3.1.3
pytesseract==0.3.13
pillow==11.0.0
sentence-transformers==2.5.0
pdf2image>=1.16.3
pytest-cov==4.1.0
numpy>=1.25.0